single test binary similar to other workspaces
There's an interesting caveat to this that main.rs needs to live inside a folder. It cannot be a plain tests/main.rs for imports to work.
