+++ reward_index/tests/integrations/main.rs
@@ -1,0 +1,4 @@
+mod common;
+
+mod iot;
+mod mobile;
